#2 Add friends likes

+++ src/main/java/ru/yandex/practicum/filmorate/dto/DtoFilm.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.dto;
 
 import lombok.Getter;
 import lombok.Setter;
-import ru.yandex.practicum.filmorate.constant.FilmCheckDate;
 import ru.yandex.practicum.filmorate.validation.FilmDataChecker;
 
 import javax.validation.constraints.Min;
@@ -13,7 +12,7 @@ import java.time.LocalDate;
 @Getter
 @Setter
 public class DtoFilm {
-    private int id;
+    private long id;
     @NotBlank(message = "name should not be blank")
     private String name;
     @Size(max = 200, message = "Description length must be less then 200")
